refactor(db-client-ts): group async-operation ops under   25_async_operation_worker_v1 (#130)
  Split async-operation.ts by concern and collect the resulting
  files under a
  new src/25_async_operation_worker_v1/ folder, mirroring the
  load-order
  directory scheme used for packages/database-src schemas:

  - asyncOperationCtl.ts — create/list dispatch + list meta
  (was async-operation.ts)
  - asyncOperationMeta.ts — loadAsyncOperation
  - asyncOperationScheduler.ts —
  asyncOperationScheduleNextPending
  - asyncOperationHelper.ts — checkRegistryForAsyncActors,
    checkRegistryAndWorkingForAsyncActors, and the AsyncActor/
    CheckRegistry*Result types they own
  - asyncOperationWorkerlet.ts — async_operation_workerlet
  table ops (was
    async-operation-workerlet.ts)

  Update index.ts export paths and the package CLAUDE.md file
  listing to match.
