+++ packages/fsm-core-db-ts/src/25_async_operation_worker_v1/asyncOperationWorkerlet.ts
@@ -1,6 +1,6 @@
 import { getLogger } from "@logtape/logtape";
-import type { DBDeps } from "./custom.types.ts";
-import { FSM_SCHEMA } from "./const.ts";
+import type { DBDeps } from "../custom.types.ts";
+import { FSM_SCHEMA } from "../const.ts";
 
 const logger = getLogger(["@pgfsm/db", "async-op-workerlet"]);
